[SPELP] : toujours les temps ...

+++ spelp/trunk/wizard/src/wilos/presentation/assistant/view/panels/ListenerTime.java
@@ -1,5 +1,5 @@
 package wilos.presentation.assistant.view.panels;
 
 public interface ListenerTime {
-	public void putValue (float val) ;
+	public void putValue (long val) ;
 }
